+++ CHANGES.md
@@ -295,3 +295,6 @@
 ### v2025.08.18
 - Self Update Checker: prepare for portable release by showing all available release files
 - Portable release: bundled with [python-3.13.7-embed-win32](https://www.python.org/ftp/python/3.13.7/python-3.13.7-embed-win32.zip) package and necessary modules
+
+### Pre-release (work-in-progress)
+- ReSound/Beltone/Interton: display full version numbers